FHIR-46524 - _typeFilter updates
(1) Remove the use of , for logical OR and replace it with repeating the _typeFilter parameter. This is currently permitted by the parameter's 0..* cardinality, but does not have any documentation around the expected behavior.

(2) Add an approach to documenting search parameters in the capability statement that are supported by the server in other contexts, but not for _typeFilter in bulk export.

(3) Clarifications on which search result parameters are supported in _typeFilter.

(4) Document bulk export's logical processing model.